refactor: NumCalculator package명 변경
