+++ src/test/java/calculator/domain/mainprocess/NumCalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator.domain.numCalculator;
+package calculator.domain.mainprocess;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
